Split LSPS2 in client/server handler modules
