+++ src/events.rs
@@ -78,8 +78,10 @@ impl EventQueue {
 /// An event which you should probably take some action in response to.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Event {
-	/// An LSPS2 (JIT Channel) protocol event.
-	LSPS2(lsps2::event::LSPS2Event),
+	/// An LSPS2 (JIT Channel) client event.
+	LSPS2Client(lsps2::event::LSPS2ClientEvent),
+	/// An LSPS2 (JIT Channel) server event.
+	LSPS2Service(lsps2::event::LSPS2ServiceEvent),
 	/// An LSPS1 protocol event.
 	#[cfg(lsps1)]
 	LSPS1(lsps1::event::Event),
