+++ src/lsps2/event.rs
@@ -15,43 +15,21 @@ use crate::prelude::{String, Vec};
 
 use bitcoin::secp256k1::PublicKey;
 
-/// An event which you should probably take some action in response to.
+/// An event which an LSPS2 client should take some action in response to.
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub enum LSPS2Event {
-	/// A request from a client for information about JIT Channel parameters.
-	///
-	/// You must calculate the parameters for this client and pass them to
-	/// [`LSPS2MessageHandler::opening_fee_params_generated`].
-	///
-	/// If an unrecognized or stale token is provided you can use
-	/// `[LSPS2MessageHandler::invalid_token_provided`] to error the request.
-	///
-	/// [`LSPS2MessageHandler::opening_fee_params_generated`]: crate::lsps2::message_handler::LSPS2MessageHandler::opening_fee_params_generated
-	/// [`LSPS2MessageHandler::invalid_token_provided`]: crate::lsps2::message_handler::LSPS2MessageHandler::invalid_token_provided
-	GetInfo {
-		/// An identifier that must be passed to [`LSPS2MessageHandler::opening_fee_params_generated`].
-		///
-		/// [`LSPS2MessageHandler::opening_fee_params_generated`]: crate::lsps2::message_handler::LSPS2MessageHandler::opening_fee_params_generated
-		request_id: RequestId,
-		/// The node id of the client making the information request.
-		counterparty_node_id: PublicKey,
-		/// The protocol version they would like to use.
-		version: u16,
-		/// An optional token that can be used as an API key, coupon code, etc.
-		token: Option<String>,
-	},
+pub enum LSPS2ClientEvent {
 	/// Information from the LSP about their current fee rates and channel parameters.
 	///
-	/// You must call [`LSPS2MessageHandler::opening_fee_params_selected`] with the fee parameter
+	/// You must call [`LSPS2ClientHandler::opening_fee_params_selected`] with the fee parameter
 	/// you want to use if you wish to proceed opening a channel.
 	///
-	/// [`LSPS2MessageHandler::opening_fee_params_selected`]: crate::lsps2::message_handler::LSPS2MessageHandler::opening_fee_params_selected
+	/// [`LSPS2ClientHandler::opening_fee_params_selected`]: crate::lsps2::client::LSPS2ClientHandler::opening_fee_params_selected
 	GetInfoResponse {
 		/// This is a randomly generated identifier used to track the JIT channel state.
 		/// It is not related in anyway to the eventual lightning channel id.
-		/// It needs to be passed to [`LSPS2MessageHandler::opening_fee_params_selected`].
+		/// It needs to be passed to [`LSPS2ClientHandler::opening_fee_params_selected`].
 		///
-		/// [`LSPS2MessageHandler::opening_fee_params_selected`]: crate::lsps2::message_handler::LSPS2MessageHandler::opening_fee_params_selected
+		/// [`LSPS2ClientHandler::opening_fee_params_selected`]: crate::lsps2::client::LSPS2ClientHandler::opening_fee_params_selected
 		jit_channel_id: u128,
 		/// The node id of the LSP that provided this response.
 		counterparty_node_id: PublicKey,
@@ -62,34 +40,10 @@ pub enum LSPS2Event {
 		min_payment_size_msat: u64,
 		/// The max payment size allowed when opening the channel.
 		max_payment_size_msat: u64,
-		/// The user_channel_id value passed in to [`LSPS2MessageHandler::create_invoice`].
+		/// The user_channel_id value passed in to [`LSPS2ClientHandler::create_invoice`].
 		///
-		/// [`LSPS2MessageHandler::create_invoice`]: crate::lsps2::message_handler::LSPS2MessageHandler::create_invoice
+		/// [`LSPS2ClientHandler::create_invoice`]: crate::lsps2::client::LSPS2ClientHandler::create_invoice
 		user_channel_id: u128,
-	},
-	/// A client has selected a opening fee parameter to use and would like to
-	/// purchase a channel with an optional initial payment size.
-	///
-	/// If `payment_size_msat` is [`Option::Some`] then the payer is allowed to use MPP.
-	/// If `payment_size_msat` is [`Option::None`] then the payer cannot use MPP.
-	///
-	/// You must generate an scid and `cltv_expiry_delta` for them to use
-	/// and call [`LSPS2MessageHandler::invoice_parameters_generated`].
-	///
-	/// [`LSPS2MessageHandler::invoice_parameters_generated`]: crate::lsps2::message_handler::LSPS2MessageHandler::invoice_parameters_generated
-	BuyRequest {
-		/// An identifier that must be passed into [`LSPS2MessageHandler::invoice_parameters_generated`].
-		///
-		/// [`LSPS2MessageHandler::invoice_parameters_generated`]: crate::lsps2::message_handler::LSPS2MessageHandler::invoice_parameters_generated
-		request_id: RequestId,
-		/// The client node id that is making this request.
-		counterparty_node_id: PublicKey,
-		/// The version of the protocol they would like to use.
-		version: u16,
-		/// The channel parameters they have selected.
-		opening_fee_params: OpeningFeeParams,
-		/// The size of the initial payment they would like to receive.
-		payment_size_msat: Option<u64>,
 	},
 	/// Use the provided fields to generate an invoice and give to payer.
 	///
@@ -106,10 +60,61 @@ pub enum LSPS2Event {
 		payment_size_msat: Option<u64>,
 		/// The trust model the LSP expects.
 		client_trusts_lsp: bool,
-		/// The `user_channel_id` value passed in to [`LSPS2MessageHandler::create_invoice`].
+		/// The `user_channel_id` value passed in to [`LSPS2ClientHandler::create_invoice`].
 		///
-		/// [`LSPS2MessageHandler::create_invoice`]: crate::lsps2::message_handler::LSPS2MessageHandler::create_invoice
+		/// [`LSPS2ClientHandler::create_invoice`]: crate::lsps2::client::LSPS2ClientHandler::create_invoice
 		user_channel_id: u128,
+	},
+}
+
+/// An event which an LSPS2 server should take some action in response to.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum LSPS2ServiceEvent {
+	/// A request from a client for information about JIT Channel parameters.
+	///
+	/// You must calculate the parameters for this client and pass them to
+	/// [`LSPS2ServiceHandler::opening_fee_params_generated`].
+	///
+	/// If an unrecognized or stale token is provided you can use
+	/// `[LSPS2ServiceHandler::invalid_token_provided`] to error the request.
+	///
+	/// [`LSPS2ServiceHandler::opening_fee_params_generated`]: crate::lsps2::service::LSPS2ServiceHandler::opening_fee_params_generated
+	/// [`LSPS2ServiceHandler::invalid_token_provided`]: crate::lsps2::service::LSPS2ServiceHandler::invalid_token_provided
+	GetInfo {
+		/// An identifier that must be passed to [`LSPS2ServiceHandler::opening_fee_params_generated`].
+		///
+		/// [`LSPS2ServiceHandler::opening_fee_params_generated`]: crate::lsps2::service::LSPS2ServiceHandler::opening_fee_params_generated
+		request_id: RequestId,
+		/// The node id of the client making the information request.
+		counterparty_node_id: PublicKey,
+		/// The protocol version they would like to use.
+		version: u16,
+		/// An optional token that can be used as an API key, coupon code, etc.
+		token: Option<String>,
+	},
+	/// A client has selected a opening fee parameter to use and would like to
+	/// purchase a channel with an optional initial payment size.
+	///
+	/// If `payment_size_msat` is [`Option::Some`] then the payer is allowed to use MPP.
+	/// If `payment_size_msat` is [`Option::None`] then the payer cannot use MPP.
+	///
+	/// You must generate an scid and `cltv_expiry_delta` for them to use
+	/// and call [`LSPS2ServiceHandler::invoice_parameters_generated`].
+	///
+	/// [`LSPS2ServiceHandler::invoice_parameters_generated`]: crate::lsps2::service::LSPS2ServiceHandler::invoice_parameters_generated
+	BuyRequest {
+		/// An identifier that must be passed into [`LSPS2ServiceHandler::invoice_parameters_generated`].
+		///
+		/// [`LSPS2ServiceHandler::invoice_parameters_generated`]: crate::lsps2::service::LSPS2ServiceHandler::invoice_parameters_generated
+		request_id: RequestId,
+		/// The client node id that is making this request.
+		counterparty_node_id: PublicKey,
+		/// The version of the protocol they would like to use.
+		version: u16,
+		/// The channel parameters they have selected.
+		opening_fee_params: OpeningFeeParams,
+		/// The size of the initial payment they would like to receive.
+		payment_size_msat: Option<u64>,
 	},
 	/// You should open a channel using [`ChannelManager::create_channel`].
 	///
